댓글삭제권한이 있을 때만 삭제

+++ src/main/java/com/sbs/jhs/at/controller/ArticleController.java
@@ -86,7 +86,14 @@ public class ArticleController {
 
 	@RequestMapping("/usr/article/doDeleteReplyAjax")
 	@ResponseBody
-	public ResultData doDeleteReplyAjax(int id) {
+	public ResultData doDeleteReplyAjax(int id, HttpServletRequest req) {
+		Member loginedMember = (Member) req.getAttribute("loginedMember");
+		ArticleReply articleReply = articleService.getForPrintArticleReplyById(id);
+
+		if (articleService.actorCanDelete(loginedMember, articleReply) == false) {
+			return new ResultData("F-1", String.format("%d번 댓글을 삭제할 권한이 없습니다.", id));
+		}
+
 		articleService.deleteReply(id);
 
 		return new ResultData("S-1", String.format("%d번 댓글을 삭제하였습니다.", id));
@@ -98,10 +105,10 @@ public class ArticleController {
 		Member loginedMember = (Member) req.getAttribute("loginedMember");
 		ArticleReply articleReply = articleService.getForPrintArticleReplyById(id);
 
-		if ( articleService.actorCanModify(loginedMember, articleReply) == false ) {
+		if (articleService.actorCanModify(loginedMember, articleReply) == false) {
 			return new ResultData("F-1", String.format("%d번 댓글을 수정할 권한이 없습니다.", id));
 		}
-		
+
 		Map<String, Object> modfiyReplyParam = Util.getNewMapOf(param, "id", "body");
 		ResultData rd = articleService.modfiyReply(modfiyReplyParam);
